==> 01_Data_concated.ipynb <==
import pandas as pd
import numpy as np

TRUYỀN DỮ LIỆU VÀO PANDAS

file_path = [r'Datasets/Raw_data/Area_population_density.xlsx',
             r'Datasets/Raw_data/Average_population_by_sex_urban_rural_residence.xlsx',
             r'Datasets/Raw_data/Crude_birth_death_rate_natural_growth_rate_locality.xlsx',
             r'Datasets/Raw_data/Immigration_emigration_net_migration_rates_locality.xlsx',
             r'Datasets/Raw_data/Sex_ratio_locality.xlsx',
             r'Datasets/Raw_data/Unemployed_data.xlsx']


df1 = pd.read_excel(file_path[0],engine='openpyxl')
df2 = pd.read_excel(file_path[1],engine='openpyxl')
df3 = pd.read_excel(file_path[2],engine='openpyxl')
df4 = pd.read_excel(file_path[3],engine='openpyxl')
df5 = pd.read_excel(file_path[4],engine='openpyxl')
df6 = pd.read_excel(file_path[5],engine='openpyxl')


THÊM NHỮNG CHỖ BỊ THIẾU TRONG CỘT VÙNG CỦA CÁC DATAFRAME

df1 = df1.ffill()
df2 = df2.ffill()
df3 = df3.ffill()
df4 = df4.ffill()
df5 = df5.ffill()
df6 = df6.ffill()

VÌ CÁC DỮ LIỆU CÓ CHỨA TIẾNG VIỆT NÊN SẼ THỰC HIỆN CHUẨN HÓA UNICODE TRƯỚC KHI CHẠY

from Vietnamese_text_normalize import normalize_vi_df,debug_unicode

df1 = normalize_vi_df(df1)
df2 = normalize_vi_df(df2)
df3 = normalize_vi_df(df3)
df4 = normalize_vi_df(df4)
df5 = normalize_vi_df(df5)
df6 = normalize_vi_df(df6)


CÁC CỘT CỦA CÁC DATAFRAME

df1.columns

df2.columns

df3.columns

df4.columns

df5.columns

df6.columns

KẾT HỢP CÁC DATAFRAME LẠI VỚI NHAU DỰA TRÊN CỘT "VÙNG" VÀ "NĂM"

data12 = normalize_vi_df(pd.merge(df1,df2, on=['Vùng','Năm']))
data34 = normalize_vi_df(pd.merge(df3,df4, on=['Vùng','Năm']))
data56 = normalize_vi_df(pd.merge(df5,df6, on=['Vùng','Năm']))

semi_data = pd.merge(data12,data34, on=['Vùng','Năm'])

semi_data


full_data = pd.merge(semi_data,data56, on=['Vùng','Năm'])

semi_data

ĐỔI TÊN MỘT SỐ CỘT BỊ TRÙNG

full_data.columns

full_data = full_data.rename(columns={'Tổng số_x':'Tổng dân số',  
                                      'Tổng số_y':'Tỉ lệ thất nghiệp',  
                                      'Thành thị':'Tỉ lệ thất nghiệp ở thành thị',
                                      'Nông thôn':'Tỉ lệ thất nghiệp ở nông thôn',
                                      'Tỷ lệ nam nữ': 'Tỷ số giới tính'
                                    })

LOẠI BỎ DỮ LIỆU CÓ VÙNG LÀ CẢ NƯỚC VÀ CÁC KHU VỰC TRONG DỮ LIỆU CHỈ ĐỂ LẠI DỮ LIỆU VỀ 63 TỈNH THÀNH TỪ NĂM 2011-2024

all = full_data[full_data['Vùng'] == 'CẢ NƯỚC'][['Năm','Tổng dân số']]

eliminate = ['CẢ NƯỚC',
             'Đồng bằng sông Hồng',
             'Trung du và miền núi phía Bắc',
             'Bắc Trung Bộ và duyên hải miền Trung',
             'Tây Nguyên',
             'Đông Nam Bộ',
             'Đồng bằng sông Cửu Long']

for region in eliminate:
    full_data = full_data[full_data['Vùng'] != region ]

LƯU DATA ĐÃ KẾT HỢP VÀ CHỈ CÒN LẠI 63 TỈNH THÀNH 

percent = full_data.merge(all,on=['Năm'],suffixes=('','_toan_quoc'))


percent.columns

percent['Phần trăm dân số'] = round((percent['Tổng dân số'] / percent['Tổng dân số_toan_quoc'])*100,2)

full_data = percent.copy()

full_data = full_data.drop(columns=['Tổng dân số_toan_quoc'])

full_data.to_csv('Datasets/Clean_data/Locality_Analyst.csv',index=False,encoding='utf-8-sig')
