==> 04_Pandas_query.ipynb <==
import pandas as pd

IMPORT DỮ LIỆU

Locality_data = pd.read_csv('Datasets/Clean_data/Locality_Analyst.csv')

Locality_data

#1. Quan sát xu hướng tăng trưởng dân số theo thời gian.
Locality_data[['Vùng', 'Năm', 'Dân số trung bình (Nghìn người)']].copy()

#2. Tìm vùng có mật độ dân số lớn nhất mỗi năm.
dt2 = Locality_data[['Vùng', 'Năm','Mật độ dân số (Người/km2)']].copy()
idx = dt2.groupby('Năm')['Mật độ dân số (Người/km2)'].idxmax()
dt2_max = dt2.loc[idx]
dt2_max


#3. Xem vùng nào có tốc độ tăng dân số tự nhiên cao/thấp nhất.
idx = Locality_data.groupby('Vùng')['Tỷ lệ tăng tự nhiên'].mean()
idx.loc[[idx.idxmax(), idx.idxmin()]]

# 4.Kiểm tra sự cân bằng nhân khẩu học
Locality_data['Tỷ suất chết thô'] = Locality_data['Tỷ suất chết thô'] * (-1)
Locality_data[['Tỷ suất sinh thô' ,'Tỷ suất chết thô']].corr(method='pearson')

#5.Đánh giá quá trình đô thị hóa của các vùng.
dt3 = Locality_data[['Vùng', 'Năm', 'Tổng dân số', 'Tổng dân số thành thị']].copy()
dt3['Tỷ lệ dân số thành thị(%)'] = round(dt3['Tổng dân số thành thị'] / dt3['Tổng dân số'] * 100, 2)
dt3[['Vùng', 'Năm','Tỷ lệ dân số thành thị(%)']]

#6. So sánh thất nghiệp giữa thành thị và nông thôn.
dt4 = Locality_data.copy()
dt4 = dt4.groupby(['Vùng'])[['Tỉ lệ thất nghiệp', 'Tỉ lệ thất nghiệp ở thành thị', 'Tỉ lệ thất nghiệp ở nông thôn']].mean().round(2)
dt4

#7. Xem vùng có tăng dân số mạnh có tỷ lệ thất nghiệp cao không.
Locality_data['Tăng dân số(%)'] = Locality_data.groupby('Vùng')['Tổng dân số'].pct_change()*100
Locality_data.groupby('Vùng')[['Tăng dân số(%)','Tỉ lệ thất nghiệp']].mean().round(2).sort_values(by='Tăng dân số(%)', ascending=False)

#8. Xem vùng nào thu hút hay mất dân cư.
dt5 = Locality_data[['Vùng', 'Tỷ suất nhập cư', 'Tỷ suất xuất cư', 'Tỷ suất di cư thuần']].copy()
dt5.groupby('Vùng')[['Tỷ suất nhập cư', 'Tỷ suất xuất cư', 'Tỷ suất di cư thuần']].mean().round(3).sort_values(ascending=True,by=[
'Tỷ suất di cư thuần'
])
