==> 07_predict.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from statsmodels.tsa.arima.model import ARIMA

Data = pd.read_csv('Datasets/Clean_data/Locality_Analyst.csv')

class ARIMAModel:
    def __init__(self,region_name):
        self.region = region_name
        
    def get_data(self):
        if self.region == 'All': return Data[['Năm','Tổng dân số']]
        
        df =  Data[Data['Vùng'] == self.region][['Năm','Tổng dân số']]
        df['Năm'] = pd.to_datetime(df['Năm'],format='%Y')
        return df
    
    def train_ARIMA_model(self,data_end_year,predict_year):
        Series = self.get_data()
        Series = Series.set_index('Năm')
        n_steps = predict_year - data_end_year
        model = ARIMA(Series,order=(1,1,1))
        model_fit = model.fit()
        forecast = model_fit.forecast(steps=n_steps)
        forecast.index = pd.date_range(
            start=Series.index[-1] + pd.DateOffset(years=1),
            periods=n_steps,
            freq='Y'
        )
        self.Draw_predict(data_end_year,predict_year,Series,forecast)
        return pd.DataFrame(forecast.values,columns=['y_pred']).to_csv(f'Result/Test_set/ARIMAModel/{self.region}_predict_result.csv')
    
    def Draw_predict(self,year,end_year,Series,Forecast):
        plt.figure(figsize=(8,4))
        plt.plot(Series.index,Series['Tổng dân số'],label='Thực tế',marker='o',linestyle = '-' )
        plt.plot(Forecast.index,Forecast,label='Dự báo',marker='o',linestyle='--',color = 'red')
        plt.title(f'Dự báo dân số {self.region} từ năm {year} - {end_year}')
        
        plt.legend()
        plt.savefig(f'Result/Test_set/ARIMAModel/Predict_plot/Dự báo dân số {self.region} từ năm {year} - {end_year}')
        plt.close()

ARIMAModel = ARIMAModel('Hà Nội')

ARIMAModel.train_ARIMA_model(data_end_year=2024,predict_year=2030)
